+++ frontend/src/components/Button/Button.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Button = ({ type, disabled, children, ...rest }) => {
+  return (
+    <button type={type} disabled={disabled} {...rest}>
+      {children}
+    </button>
+  );
+};
+export { Button };
