+++ frontend/src/components/Button/index.js
@@ -1,10 +1,1 @@
-import React from 'react';
-
-const Button = ({ type, disabled, children, ...rest }) => {
-  return (
-    <button type={type} disabled={disabled} {...rest}>
-      {children}
-    </button>
-  );
-};
-export default Button;
+export { Button } from './Button';
